fix(cluster_k8s): rename minikube.py to mini_k8s.py
